PJAPI/requirements.txt
django
djangorestframework
django-cors-headers
sqlparse